backend/requirements.dev.txt
Django>=4.1,<5.0
djangorestframework>=3.14
psycopg2>=2.9
pylint>=2.16
pylint-django>=2.5
